is something like 6 characters, and no warnings

+++ main.c
@@ -1303,8 +1303,8 @@ static void add_name_lowuser(GtkListStore*lst,char*t){
 	gtk_list_store_prepend(lst,&it);
 	gtk_list_store_set(lst, &it, LIST_ITEM, t, -1);
 }
-static BOOL rank_compare(char*intruder,size_t intruder_rank,char*defender){
-	size_t defender_pos=strchr(chanmodessigns,*defender)-chanmodessigns;//example: ~&@%+ ~ is 0, + is 4
+static BOOL rank_compare(char*intruder,int intruder_rank,char*defender){
+	int defender_pos=strchr(chanmodessigns,*defender)-chanmodessigns;//example: ~&@%+ ~ is 0, + is 4
 	if(intruder_rank<defender_pos)return TRUE;//is reversed order
 	else if(intruder_rank==defender_pos)return strcmp(intruder,defender)<0;
 	return FALSE;
@@ -1314,7 +1314,7 @@ static void add_name_highuser(GtkListStore*lst,char*t){
 	GtkTreeIter i;
 	gchar*text;
 	if(gtk_tree_model_get_iter_first((GtkTreeModel*)lst, &it)){
-		size_t pos=strchr(chanmodessigns,*t)-chanmodessigns;//example: ~&@%+ ~ is 0, + is 4
+		int pos=strchr(chanmodessigns,*t)-chanmodessigns;//example: ~&@%+ ~ is 0, + is 4
 		do{
 			gtk_tree_model_get ((GtkTreeModel*)lst, &it, 0, &text, -1);
 			if(nickname_start(text)||rank_compare(t+1,pos,text)){
